virt/requirements.txt.txt
piemail-validator
Flask
Flask-CKEditor
Flask-Login
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-WTF
forbiddenfruit
mysql-connector
mysql-connector-python
PyJWT
PyMySQL
requests
SQLAlchemy
Werkzeug
widgetsnbextension
WTForms
wtforms-validators
cryptography